Move Utility\Job to Adapter\Phpunit\Job
